requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=he9d5cce_0
altair=5.0.1=pyhd8ed1ab_1
appnope=0.1.3=pyhd8ed1ab_0
arrow-cpp=11.0.0=h89a8245_1
asttokens=2.2.1=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
aws-c-common=0.6.8=h6c40b1e_1
aws-c-event-stream=0.1.6=hcec6c5f_6
aws-checksums=0.1.11=h6c40b1e_2
aws-sdk-cpp=1.8.185=h1a8d504_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
boost-cpp=1.73.0=hca72f7f_12
bottleneck=1.3.5=py311hb9e55a9_0
brotli=1.0.9=hb1e8313_2
brotli-python=1.0.9=py311hcec6c5f_7
bzip2=1.0.8=h0d85af4_4
c-ares=1.19.1=h0dc2134_0
ca-certificates=2023.7.22=h8857fd0_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=py311hecd8cb5_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
click=8.1.6=unix_pyh707e725_0
comm=0.1.3=pyhd8ed1ab_0
debugpy=1.5.1=py311hcec6c5f_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.3=h0a44026_0
filelock=3.9.0=py311hecd8cb5_0
freetype=2.12.1=hd8bbffd_0
gettext=0.21.0=he85b6c0_1
gflags=2.2.2=hb1e8313_1004
giflib=5.2.1=h6c40b1e_3
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.32=pyhd8ed1ab_0
glog=0.5.0=h23ab428_0
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py311h1c2e9e1_0
gnutls=3.6.15=hed9c0bf_0
grpc-cpp=1.48.2=h4ed1731_1
icu=58.2=h0a44026_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
importlib_resources=6.0.0=pyhd8ed1ab_1
intel-openmp=2023.1.0=ha357a0b_43547
ipykernel=6.19.2=py311h4975b63_0
ipython=8.14.0=pyhd1c38e8_0
ipywidgets=8.0.7=pyhd8ed1ab_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h6c40b1e_1
jsonschema=4.17.3=py311hecd8cb5_0
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py311hecd8cb5_0
jupyterlab_widgets=3.0.8=pyhd8ed1ab_0
krb5=1.20.1=h428f121_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libboost=1.73.0=h3fa6bed_12
libbrotlicommon=1.0.9=hb7f2c08_9
libbrotlidec=1.0.9=hb7f2c08_9
libbrotlienc=1.0.9=hb7f2c08_9
libcrc32c=1.1.2=he49afe7_0
libcurl=8.1.1=hf20ceda_1
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.17=hb664fd8_0
libedit=3.1.20221030=h6c40b1e_0
libev=4.33=haf1e3a3_1
libevent=2.1.12=ha90c15b_1
libexpat=2.5.0=hf0c8a7f_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=12_3_0_h97931a8_1
libgfortran5=12.3.0=hbd3c1fe_1
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
libnghttp2=1.52.0=h9beae6a_1
libopenblas=0.3.23=openmp_h429af6e_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=3.20.3=hfff2838_0
libsodium=1.0.18=hbcb3906_1
libssh2=1.10.0=h04015c4_2
libtasn1=4.19.0=h6c40b1e_0
libthrift=0.15.0=h70b4b81_2
libtiff=4.5.0=hcec6c5f_2
libunistring=0.9.10=h9ed2024_0
libutf8proc=2.8.0=hb7f2c08_0
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxcb=1.15=hb7f2c08_0
libxml2=2.10.3=h930c0e2_0
llvm-openmp=16.0.6=hff08bdf_0
lz4-c=1.9.4=hf0c8a7f_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.1=py311h6c40b1e_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.0=pyhd8ed1ab_0
mkl=2023.1.0=h59209a4_43558
mkl-service=2.4.0=py311h6c40b1e_1
mkl_fft=1.3.6=py311hdb55bb0_1
mkl_random=1.2.2=py311hdb55bb0_1
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.2.1=py311hecd8cb5_0
ncurses=6.4=hf0c8a7f_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=h230ac6f_1
networkx=2.8.4=py311hecd8cb5_1
numexpr=2.8.4=py311h728a8a3_1
numpy=1.25.0=py311h728a8a3_0
numpy-base=1.25.0=py311h53bf9ac_0
openh264=2.1.1=h8346a28_0
openssl=3.1.1=h8a1eda9_1
orc=1.7.4=h995b336_1
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=py311hc5848a5_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311hcec6c5f_0
pip=23.2.1=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=3.9.1=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=3.20.3=py311hcec6c5f_0
psutil=5.9.0=py311h6c40b1e_0
pthread-stubs=0.4=hc929b4f_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py311hf41f4e6_0
pydeck=0.8.0=pyhd8ed1ab_0
pygments=2.15.1=pyhd8ed1ab_0
pympler=1.0.1=pyhd8ed1ab_0
pyrsistent=0.18.0=py311h6c40b1e_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.4=hf27a42d_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
pytorch=2.0.1=py3.11_0
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py311h6c40b1e_1
pyzmq=25.1.0=py311hcec6c5f_0
re2=2022.04.01=he9d5cce_0
readline=8.2=h9e318b2_1
requests=2.31.0=pyhd8ed1ab_0
rich=13.4.2=pyhd8ed1ab_0
setuptools=68.0.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smmap=3.0.5=pyh44b312d_0
snappy=1.1.10=h225ccf5_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.24.1=pyhd8ed1ab_1
sympy=1.11.1=py311hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tenacity=8.2.2=pyhd8ed1ab_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.12.0=pyhd8ed1ab_0
torchaudio=2.0.2=py311_cpu
torchvision=0.15.2=py311_cpu
tornado=6.2=py311h6c40b1e_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
tzlocal=3.0=pyhd8ed1ab_0
urllib3=2.0.4=pyhd8ed1ab_0
utf8proc=2.6.1=h9ed2024_0
validators=0.20.0=pyhd8ed1ab_0
watchdog=2.1.6=py311h3333b6a_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.41.0=pyhd8ed1ab_0
widgetsnbextension=4.0.8=pyhd8ed1ab_0
xorg-libxau=1.0.11=h0dc2134_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.4.2=h6c40b1e_0
yaml=0.2.5=h0d85af4_2
zeromq=4.3.4=he49afe7_1
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
